Moved warmup code, and renamed functions.
commit-id:4ea33894
